+++ src/main/java/org/kumoricon/site/utility/loadbasedata/LoadBaseDataPresenter.java
@@ -123,7 +123,6 @@ public class LoadBaseDataPresenter {
             {"view_check_in_by_hour_report", "View attendee check ins per hour report"},
             {"view_staff_report", "View staff report (lists name/phone numbers)"},
             {"view_role_report", "View registration system role report"},
-            {"view_check_in_by_badge_report", "View check in time by badge type report"},
             {"manage_staff", "Add/edit users and reset passwords"},
             {"manage_pass_types", "Add/edit badge types"},
             {"manage_roles", "Add/edit security roles"},
@@ -148,17 +147,15 @@ public class LoadBaseDataPresenter {
         roles.put("Coordinator", new String[] {"at_con_registration", "pre_reg_check_in", "attendee_search",
                                                "print_badge", "attendee_edit", "attendee_add_note",
                                                "reprint_badge", "view_staff_report",
-                                               "view_check_in_by_hour_report", "view_check_in_by_badge_report"});
+                                               "view_check_in_by_hour_report"});
         roles.put("Coordinator - VIP Badges", new String[] {"at_con_registration", "pre_reg_check_in",
                                                             "attendee_search", "print_badge", "attendee_edit",
                                                             "attendee_add_note", "reprint_badge", "view_staff_report",
-                                                            "view_check_in_by_hour_report",
-                                                            "view_check_in_by_badge_report", "badge_type_vip"});
+                                                            "view_check_in_by_hour_report", "badge_type_vip"});
         roles.put("Coordinator - Other Badges", new String[] {"at_con_registration", "pre_reg_check_in",
                                                               "attendee_search", "print_badge", "attendee_edit",
                                                               "attendee_add_note", "reprint_badge", "view_staff_report",
-                                                              "view_check_in_by_hour_report",
-                                                              "view_check_in_by_badge_report", "badge_type_artist",
+                                                              "view_check_in_by_hour_report", "badge_type_artist",
                                                               "badge_type_exhibitor", "badge_type_guest",
                                                               "badge_type_panelist", "badge_type_industry"});
         roles.put("Manager", new String[] {"at_con_registration", "pre_reg_check_in", "attendee_search",
@@ -166,16 +163,16 @@ public class LoadBaseDataPresenter {
                 "badge_type_vip", "badge_type_press", "badge_type_artist", "badge_type_exhibitor", "badge_type_guest",
                 "badge_type_industry", "badge_type_panelist", "badge_type_staff",
                 "attendee_override_price", "reprint_badge", "manage_staff", "view_staff_report", "view_attendance_report",
-                "view_check_in_by_hour_report", "view_check_in_by_badge_report"});
+                "view_check_in_by_hour_report"});
         roles.put("Director", new String[] {"at_con_registration", "pre_reg_check_in", "attendee_search",
                 "print_badge", "attendee_edit", "attendee_add_note",
                 "attendee_override_price", "reprint_badge", "manage_staff", "manage_pass_types",
                 "badge_type_vip", "badge_type_press", "badge_type_artist", "badge_type_exhibitor", "badge_type_guest",
                 "badge_type_industry", "badge_type_panelist", "badge_type_staff",
                 "view_role_report",
-                "view_attendance_report", "view_attendance_report_revenue", "view_staff_report", "view_check_in_by_hour_report",
-                "view_check_in_by_badge_report"});
-        roles.put("Ops", new String[] {"attendee_search", "attendee_add_note", "view_check_in_by_badge_report"});
+                "view_attendance_report", "view_attendance_report_revenue", "view_staff_report",
+                "view_check_in_by_hour_report"});
+        roles.put("Ops", new String[] {"attendee_search", "attendee_add_note"});
 
         HashMap<String, Right> rightMap = getRightsHashMap();