+++ CHANGES.md
@@ -1,6 +1,8 @@
 Change Log
 ----------------
 - Version 0.9.7 (Pending)
+    - Removed Check In by Badge Report, added Search by Badge Type screen. (Attendees can be 
+      edited from the Search by Badge Type screen with the proper permission)
 
 - Version 0.9.6 (9/13/2016)
     - Changed error message to disappear automatically after a few seconds, instead of requiring a click